J1939: Fix float conversions
Created a function which takes in two can bytes, a conversion factor,
and an offset and returns a float value and updates all the usages in
CANInterpret to use it for decoding instead of functions to convert.
We were seeing integer overflows in the speed display function when
going over 350 RPM due to this problem. This also updates all the other
decoding cases to also use the new function.