PedalBox bringup and misc fixes (#39)
* CAN_SPEC: Pull in updated header guards

* Port Pedal Box Support

- Add new USE_APPS define to switch between pedal box + torque mode
  and the rotary potentiometer with speed mode.
- Port defines for pins and voltages from `pedalBox` branch
- Fix GetPedalTorque logic to convert the 0-100% torque value to the
  MIN/MAX TORQUE_VAL variables which maps to the correct bytes to
  send over CAN.

* TaskScheduler: UpdateSpeed -> UpdateCommandedPower

This function really just updates the command message contents that
we send to the inverter but it doesn't care about the conversion for
the format for either speed or torque. It expects that work has already
been done. The new name makes that more clear.

* scripts: break out compile_app

This wraps the call to arduino-cli from flash_arduino.sh into its
own script that can be called on its own.

* FreeRTOSVariant.h: Fix spacing

* Pedal: Improvements from testing

AppConfig.h:
 - Update GPIO's for potentiometer vs APPS
 - Update thresholds for left/right APPs pots
 - Improve mapping logic for min/max speed/torque to better match
   the CAN spec.
 - Update max torque percentage to match testing in ECE109
 - Split DISPLAYS_ENABLED into two different defines for the LCD and
   the 7-segment displays
 - Replace coolant temp display with average torque percent display

InverterApp.ino:
 - Rename TaskInverterStateMachineControl to
   TaskPD400InverterStateMachineControl as we will eventually add a
   separate state machine control algo for the pythagoras inverter.
 - Update task priorities for machine control and the LCDs
 - Update task delays for displays tasks
 - Revert fault logic to the freeRTOS task instead of machine control
 - Fix infinite printing of the "Controlled Power Down" and "Fail Safe
   State" messages to the serial console

* J1939: Fix float conversions

Created a function which takes in two can bytes, a conversion factor,
and an offset and returns a float value and updates all the usages in
CANInterpret to use it for decoding instead of functions to convert.
We were seeing integer overflows in the speed display function when
going over 350 RPM due to this problem. This also updates all the other
decoding cases to also use the new function.

* J1939: Split CanInterpret

This moves some of the more complicated logic into smaller, more
readable helper functions.